Renamed to pg_hibernator and updated to work with Postgres 9.4.
Upon a couple of requests, to avoid name collision with the popular ORM
Hibernate, I decided to change the name of the extension to
Postgres Hibernator.

Updated the README, Makefile and code to use the new name everywhere. I
decided against changing the name of the C file so that the development
history does not get lost in file renames. To accomodate for a C
filename the is different from compiled object's name, I had to use the
MODULE_big and OBJS variables in Makefile.

Also updated the code to use Postgres 9.4's updated infrastructure for
buffer-manager partitioning, while retaining compatibility with Postgres
version 9.3.