Update README to include description of parameters.
Also removed a caveat, since it doesn't apply anymore as both the pre-
and post-9.4 versions of the Postgres Hibernator restore one database at
a time, by default. And the decription of pg_hobernator.parallel
perameter covers the case when it may help/hurt to enable parallel
restores of databases.